refactor(e2e): address review — generic screenshot ignore, name the offline helper
Two review comments from @enahum that were straightforwardly right.

.gitignore listed four Maestro screenshots by name. The name is just the flow's
own takeScreenshot label, so the list grows with every flow that captures one.
Replaced with detox/maestro/*.png. No PNG under detox/maestro is tracked, so the
wildcard hides nothing that was being kept.

support/utils/network.ts is not general network plumbing: its whole surface is
isNetworkControlAvailable / goOffline / goOnline, i.e. simulating no
connectivity. Renamed to offline_simulation.ts so the filename says so. It is
consumed through the utils barrel, so the rename touches only that re-export
plus one doc reference in pending_posts_offline.e2e.ts.
