+++ .gitignore
@@ -153,7 +153,6 @@ detox/maestro/.maestro-test-env*.sh
 detox/maestro/.maestro-test-env*.bak
 
 .e2e-debug/
-detox/maestro/report-problem-action-opened.png
-detox/maestro/report-link-browser-opened.png
-detox/maestro/report-link-settings-reachable.png
-detox/maestro/report-mail-settings-reachable.png
+# Maestro takeScreenshot output when a flow is run locally without
+# --test-output-dir; the name is the flow's own label, so match any of them.
+detox/maestro/*.png
