+++ detox/e2e/support/utils/index.ts
@@ -9,7 +9,7 @@ import {logDebug} from '../../../provision/log';
 
 export * from './email';
 export * from './detoxhelpers';
-export * from './network';
+export * from './offline_simulation';
 
 export const wait = async (ms: number): Promise<any> => {
     return new Promise((resolve) => setTimeout(resolve, ms));
